Move LocalStorage to pkg/storage (with ListStorage)
