+++ pkg/storage/list.go
@@ -5,8 +5,6 @@ import (
 	"github.com/devfile/library/pkg/devfile/generator"
 	"github.com/devfile/library/pkg/devfile/parser"
 	"github.com/devfile/library/pkg/devfile/parser/data/v2/common"
-
-	"github.com/redhat-developer/odo/pkg/localConfigProvider"
 )
 
 const (
@@ -14,9 +12,23 @@ const (
 	DefaultVolumeSize = "1Gi"
 )
 
+// LocalStorage holds storage related information
+type LocalStorage struct {
+	// Name of the storage
+	Name string `yaml:"Name,omitempty"`
+	// Size of the storage
+	Size string `yaml:"Size,omitempty"`
+	// Boolean indicating if the volume should be ephemeral. A nil pointer indicates to use the default behaviour
+	Ephemeral *bool `yaml:"Ephemeral,omitempty"`
+	// Path of the storage to which it will be mounted on the container
+	Path string `yaml:"Path,omitempty"`
+	// Container is the container name on which this storage is mounted
+	Container string `yaml:"-" json:"-"`
+}
+
 // ListStorage gets all the storage from the devfile.yaml
-func ListStorage(devfileObj parser.DevfileObj) ([]localConfigProvider.LocalStorage, error) {
-	var storageList []localConfigProvider.LocalStorage
+func ListStorage(devfileObj parser.DevfileObj) ([]LocalStorage, error) {
+	var storageList []LocalStorage
 
 	volumeMap := make(map[string]devfilev1.Volume)
 	components, err := devfileObj.Data.GetComponents(common.DevfileOptions{})
@@ -41,7 +53,7 @@ func ListStorage(devfileObj parser.DevfileObj) ([]localConfigProvider.LocalStora
 		for _, volumeMount := range component.Container.VolumeMounts {
 			vol, ok := volumeMap[volumeMount.Name]
 			if ok {
-				storageList = append(storageList, localConfigProvider.LocalStorage{
+				storageList = append(storageList, LocalStorage{
 					Name:      volumeMount.Name,
 					Size:      vol.Size,
 					Ephemeral: vol.Ephemeral,
